+++ ui/.dir-locals.el
@@ -1,0 +1,1 @@
+((nil . ((create-lockfiles . nil))))
