+++ ui/src/proxy/forms/binaries.js
@@ -1,0 +1,140 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+import exact from 'prop-types-exact';
+import {connect} from 'react-redux';
+
+import {CustomInput, Col, Row} from 'reactstrap';
+
+import {
+    PlaygroundForm, PlaygroundFormGroup,
+    PlaygroundFormGroupRow} from '../../shared/forms';
+
+import BinaryIcon from '../../app/images/binary.png';
+import {updateForm} from '../../app/store';
+import {readFile} from '../../shared/utils';
+import {ActionRemove} from '../../shared/actions';
+
+
+const mapStateToProps = function(state, other) {
+    return {
+        form: state.form.value,
+        proxies: state.proxy.value,
+        meta: state.meta.value,
+    };
+}
+
+
+export class BinariesListForm extends React.PureComponent {
+    static propTypes = exact({
+        onDelete: PropTypes.func.isRequired,
+        binaries: PropTypes.object.isRequired,
+    });
+
+    render () {
+        const {onDelete, binaries} = this.props;
+        const title = '';
+
+        if (Object.keys(binaries).length === 0) {
+            return '';
+        }
+        return (
+            <Row className="mt-2 pb-3">
+              <Col>
+                <Row className="pl-5 pr-5">
+                  <Col sm={1} className="m-0 p-0">
+                    <div className="p-1 bg-dark">
+                      <span>&nbsp;</span>
+                    </div>
+                  </Col>
+                  <Col sm={11} className="m-0 p-0">
+                    <div className="p-1 pl-4 bg-dark">
+                      Path
+                    </div>
+                  </Col>
+                </Row>
+                {Object.keys(binaries).map((name, index) => {
+                    return (
+                        <Row key={index} className="pl-5 pr-5">
+                          <Col sm={1} className="m-0 p-0">
+                            <div className="p-2 bg-white border-bottom">
+                              <ActionRemove
+                                title={title}
+                                name={title}
+                                remove={evt => onDelete(name)} />
+                            </div>
+                          </Col>
+                          <Col sm={11} className="m-0 p-0 border-bottom bg-white">
+                              <img
+                                alt={name}
+                                src={BinaryIcon}
+                                width="18px"
+                                className="m-2 ml-1 mr-2"  />
+                            binary/{name}
+                          </Col>
+                        </Row>);
+                })}
+              </Col>
+            </Row>);
+    }
+}
+
+
+export class BaseProxyBinariesForm extends React.PureComponent {
+    static propTypes = exact({
+        dispatch: PropTypes.func.isRequired,
+        form: PropTypes.object.isRequired,
+        proxies: PropTypes.object.isRequired,
+        meta: PropTypes.object.isRequired,
+    });
+
+    onChange = async (evt) => {
+        const {dispatch, form} = this.props;
+        const {binaries={}} = form;
+        const update = {};
+        update[evt.target.files[0].name] = await readFile(evt.target.files[0]);
+        await dispatch(updateForm({binaries: {...binaries, ...update}}));
+    }
+
+    onDelete = async (name) => {
+        const {dispatch, form} = this.props;
+        const {binaries: _binaries={}} = form;
+        const binaries = {..._binaries};
+        delete binaries[name];
+        await dispatch(updateForm({binaries}));
+    }
+
+    get messages () {
+        return [
+            "Add binary files required by your proxy.",
+            "The files will be accessible to Envoy with the uploaded `filename` in the `binary/` folder."
+        ];
+    }
+
+    render () {
+        const {form} = this.props;
+        const {binaries={}} = form;
+        return (
+            <PlaygroundForm messages={this.messages}>
+              <PlaygroundFormGroup>
+                <PlaygroundFormGroupRow
+                  label="binaries"
+                  title="Add a binary">
+                  <Col sm={8}>
+                    <CustomInput
+                      type="file"
+                      onInput={this.onChange}
+                      id="binaries"
+                      name="binaries" />
+                  </Col>
+                </PlaygroundFormGroupRow>
+                <BinariesListForm
+                  onDelete={this.onDelete}
+                  binaries={{...binaries}} />
+              </PlaygroundFormGroup>
+            </PlaygroundForm>
+        );
+    }
+}
+
+const ProxyBinariesForm = connect(mapStateToProps)(BaseProxyBinariesForm);
+export {ProxyBinariesForm};
