+++ ui/src/proxy/forms/index.js
@@ -1,0 +1,5 @@
+
+export {ProxyForm} from './proxy';
+export {ProxyLoggingForm} from './logging';
+export {ProxyBinariesForm} from './binaries';
+export {ProxyCertificatesForm} from './certificates';
